Feat : Item Entity 연관 Entity 생성
