+++ src/main/java/techit/gongsimchae/domain/common/participate/entity/Participate.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import techit.gongsimchae.domain.BaseEntity;
-import techit.gongsimchae.domain.admin.item.entity.Item;
+import techit.gongsimchae.domain.groupbuying.item.entity.Item;
 import techit.gongsimchae.domain.common.user.entity.User;
 
 import java.time.LocalDateTime;
