+++ src/main/java/techit/gongsimchae/domain/groupbuying/delivery/entity/Delivery.java
@@ -1,4 +1,23 @@
 package techit.gongsimchae.domain.groupbuying.delivery.entity;
 
+import jakarta.persistence.Embedded;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import techit.gongsimchae.domain.Address;
+
+@Entity
+@Getter
+@NoArgsConstructor
 public class Delivery {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Embedded
+    private Address address;
+    private DeliveryStatus deliveryStatus;
 }
