+++ src/main/java/techit/gongsimchae/domain/groupbuying/delivery/entity/DeliveryStatus.java
@@ -1,0 +1,10 @@
+package techit.gongsimchae.domain.groupbuying.delivery.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum DeliveryStatus {
+
+    배달완료,배달중,배달시작전,취소
+
+}
