+++ src/main/java/techit/gongsimchae/domain/groupbuying/orderitem/entity/OrderItem.java
@@ -1,4 +1,36 @@
 package techit.gongsimchae.domain.groupbuying.orderitem.entity;
 
-public class OrderItem {
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import techit.gongsimchae.domain.BaseEntity;
+import techit.gongsimchae.domain.groupbuying.item.entity.Item;
+import techit.gongsimchae.domain.groupbuying.itemoption.entity.ItemOption;
+import techit.gongsimchae.domain.groupbuying.orders.entity.Orders;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class OrderItem extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private Integer price;
+    private Integer count;
+    @ManyToOne
+    @JoinColumn(name = "orders_id")
+    private Orders orders;
+    @ManyToOne
+    @JoinColumn(name = "item_id")
+    private Item item;
+    @OneToOne
+    @JoinColumn(name = "item_option_id")
+    private ItemOption itemOption;
 }
