+++ src/main/java/techit/gongsimchae/domain/groupbuying/orderitem/entity/OrderStatus.java
@@ -1,0 +1,10 @@
+package techit.gongsimchae.domain.groupbuying.orderitem.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum OrderStatus {
+
+    주문,취소,완료
+
+}
