+++ src/main/java/techit/gongsimchae/domain/groupbuying/orders/entity/Orders.java
@@ -1,4 +1,25 @@
 package techit.gongsimchae.domain.groupbuying.orders.entity;
 
-public class Orders {
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import techit.gongsimchae.domain.BaseEntity;
+import techit.gongsimchae.domain.groupbuying.orderitem.entity.OrderStatus;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Orders extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Enumerated(EnumType.ORDINAL)
+    private OrderStatus orderStatus;
 }
